Rename ts files to js for easier review
