+++ src/EventTarget.js
